fix(harvest): :nail_care: Split withKonnectorJob from the model
